+++ examples/basic/src/http.ts
@@ -1,0 +1,33 @@
+// Helper to parse JSON body safely
+export async function parseJsonBody(request: Request): Promise<any> {
+  try {
+    return await request.json();
+  } catch (error) {
+    return {};
+  }
+}
+
+// Helper for CORS headers
+export function corsHeaders() {
+  return {
+    "Access-Control-Allow-Origin": "*",
+    "Access-Control-Allow-Methods": "GET, POST, DELETE, OPTIONS",
+    "Access-Control-Allow-Headers": "Content-Type",
+  };
+}
+
+// Helper for error responses
+export function errorResponse(message: string, status: number = 400) {
+  return new Response(JSON.stringify({ error: message }), {
+    status,
+    headers: { "Content-Type": "application/json", ...corsHeaders() },
+  });
+}
+
+// Helper for success responses
+export function jsonResponse(data: any, status: number = 200) {
+  return new Response(JSON.stringify(data), {
+    status,
+    headers: { "Content-Type": "application/json", ...corsHeaders() },
+  });
+}
